v2.0.0
- Description for version numbering system added to README.md
- Functions for indexer and flywheel created
- Function for expansion adjusted for the new expansion module
- Started creating odometry & PID functions
- Created competition_intialize.hpp and competition_intialize.cpp, and moved all pre-auton exclusive functions into those files
- Removed all button boolean variables
- Added Full AWP to auton select options
- Installed sylib
- Removed flywheel.cpp, intake.cpp, and their respective header files
- Added delays after indexer movements
- Untracked cache files

+++ src/subsystemFiles/drive.cpp
@@ -1,7 +1,7 @@
 #include "main.h"
 
 float getMotorOutput(int i) {
-    return i*exp((double)((abs(i)-100)*60)/1000);
+    return i*exp((int)(((abs(i)-100)*60)/1000));
 }
 
 void DirectionToggle() {
@@ -18,26 +18,26 @@ void setDriveMotors() {
     if (abs(RYAxis) < 10) RYAxis = 0;
 
     if (Direction == 'N') {
-        /*
         BLM.move(getMotorOutput(LYAxis));
         FLM.move(getMotorOutput(LYAxis));
         BRM.move(getMotorOutput(RYAxis));
         FRM.move(getMotorOutput(RYAxis));
-        */
+        /*
         BLM.move(LYAxis);
         FLM.move(LYAxis);
         BRM.move(RYAxis);
         FRM.move(RYAxis);
+        */
     } else if (Direction == 'R') {
-        /*
         BLM.move(getMotorOutput(RYAxis) * -1);
         FLM.move(getMotorOutput(RYAxis) * -1);
         BRM.move(getMotorOutput(LYAxis) * -1);
         FRM.move(getMotorOutput(LYAxis) * -1);
-        */
+        /*
         BLM.move(RYAxis * -1);
         FLM.move(RYAxis * -1);
         BRM.move(LYAxis * -1);
         FRM.move(LYAxis * -1);   
+        */
     }
 }
